+++ src/components/nodes-table/ExpandedNode.vue
@@ -10,11 +10,12 @@
       <v-tab key="node">Node</v-tab>
       <v-tab v-if="showHass" key="homeassistant">Home Assistant</v-tab>
       <v-tab key="groups">Groups</v-tab>
-      <v-tab v-if="$vuetify.mdAndUp" key="debug">Debug Info</v-tab>
+      <v-tab v-if="$vuetify.breakpoint.mdAndUp" key="debug">Debug Info</v-tab>
 
       <!-- TABS -->
       <v-tabs-items
         style="background: transparent; padding-bottom: 10px;"
+        touchless
         v-model="currentTab"
       >
         <!-- TAB NODE -->
@@ -35,11 +36,11 @@
 
         <!-- TAB GROUPS -->
         <v-tab-item key="groups">
-          <association-groups :node="node" :nodes="nodes" :socket="socket" />
+          <association-groups :node="node" :socket="socket" />
         </v-tab-item>
 
         <!-- TAB DEBUG INFO -->
-        <v-tab-item key="debug">
+        <v-tab-item v-if="$vuetify.breakpoint.mdAndUp" key="debug">
           <v-textarea
             class="mx-2"
             rows="10"
@@ -68,7 +69,6 @@ export default {
     headers: Array,
     isMobile: Boolean,
     node: Object,
-    nodes: Array,
     socket: Object
   },
   components: {