+++ src/components/nodes-table/NodeDetails.vue
@@ -244,8 +244,6 @@ export default {
       this.$listeners.export(this.node, 'node_' + this.node.id, 'json')
     },
     getValue (v) {
-      // const node = this.nodes[v.nodeId]
-
       if (this.node && this.node.values) {
         return this.node.values.find(i => i.id === v.id)
       } else {
@@ -378,7 +376,7 @@ export default {
     },
     validateTopic (name) {
       const match = name
-        ? name.match(/[/a-zA-Z\u00C0-\u024F\u1E00-\u1EFF0-9_-]+/g)
+        ? name.match(/[/a-zA-Z\u00C0-\u024F\u1E00-\u1EFF0-9 _-]+/g)
         : [name]
 
       return match[0] !== name ? 'Only a-zA-Z0-9_- chars are allowed' : null