+++ src/router/index.js
@@ -108,6 +108,10 @@ router.beforeEach(async (to, from, next) => {
     return
   }
 
+  if (store.state.auth === undefined && to.path !== Routes.login) {
+    localStorage.setItem('nextUrl', to.path)
+  }
+
   if (store.state.auth === false) {
     if (to.path === Routes.login) {
       next({